requirements.txt
aiofiles==22.1.0
aiosmtplib==1.1.7
aiosqlite==0.17.0
anyio==3.6.2
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-mail==1.2.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
Pillow==9.3.0
pydantic==1.10.2
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
